remove outdated duplicate implementation of slice
